Move navigation blocker to role details; patch up blocker showing on save click
