+++ src/pages/daos/[daoAddress]/roles/edit/details/index.tsx
@@ -4,6 +4,7 @@ import {
   Drawer,
   DrawerBody,
   DrawerContent,
+  DrawerOverlay,
   Flex,
   Icon,
   IconButton,
@@ -15,7 +16,7 @@ import { ArrowLeft, DotsThree, Trash, X } from '@phosphor-icons/react';
 import { FieldArray, useFormikContext } from 'formik';
 import { useEffect, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useNavigate, useSearchParams } from 'react-router-dom';
+import { Blocker, useBlocker, useNavigate, useSearchParams } from 'react-router-dom';
 import { Hex, isHex } from 'viem';
 import RoleFormTabs from '../../../../../../components/pages/Roles/forms/RoleFormTabs';
 import {
@@ -25,12 +26,15 @@ import {
 } from '../../../../../../components/pages/Roles/types';
 import {
   CARD_SHADOW,
+  MOBILE_DRAWER_OVERLAY,
   NEUTRAL_2_82_TRANSPARENT,
+  SEXY_BOX_SHADOW_T_T,
   useHeaderHeight,
 } from '../../../../../../constants/common';
 import { DAO_ROUTES } from '../../../../../../constants/routes';
 import { useFractal } from '../../../../../../providers/App/AppProvider';
 import { useNetworkConfig } from '../../../../../../providers/NetworkConfig/NetworkConfigProvider';
+import { UnsavedChangesWarningContent } from '../unsavedChangesWarningContent';
 
 function EditRoleMenu({ onRemove, hatId }: { hatId: Hex; onRemove: () => void }) {
   const { t } = useTranslation(['roles']);
@@ -129,127 +133,174 @@ export default function RoleEditDetails() {
   const [searchParams] = useSearchParams();
   const hatEditingId = searchParams.get('hatId');
   const hatIndex = values.hats.findIndex(h => h.id === hatEditingId);
+
+  const blocker: Blocker = useBlocker(({ currentLocation, nextLocation }) => {
+    // console.log({
+    //   roleEditing: values.roleEditing,
+    //   currentLocation,
+    //   nextLocation,
+    // });
+
+    return !!values.roleEditing && currentLocation.pathname !== nextLocation.pathname;
+  });
+
   if (!isHex(hatEditingId)) return null;
   if (!daoAddress) return null;
   if (hatEditingId === undefined) return null;
 
+  const goBackToRolesEdit = () =>
+    navigate(DAO_ROUTES.rolesEdit.relative(addressPrefix, daoAddress), { replace: true });
+
   return (
-    <FieldArray name="hats">
-      {({ remove, push }) => (
-        <>
-          <Show below="md">
-            <Portal>
-              <Box
-                position="fixed"
-                top={0}
-                h="100vh"
-                w="full"
-                bg="neutral-1"
-                px="1rem"
-                pt={headerHeight}
-                overflow="scroll"
-              >
-                <Flex
-                  justifyContent="space-between"
-                  alignItems="center"
-                  my="1.75rem"
+    <>
+      {blocker.state === 'blocked' && (
+        // <Hide above="md">
+        <Drawer
+          placement="bottom"
+          isOpen={true}
+          onClose={() => {}}
+        >
+          <DrawerOverlay
+            bg={MOBILE_DRAWER_OVERLAY}
+            backdropFilter="blur(6px)"
+          />
+          <DrawerContent
+            bg={NEUTRAL_2_82_TRANSPARENT}
+            border="none"
+            borderTopRightRadius="1rem"
+            borderBottomRightRadius="1rem"
+            boxShadow={SEXY_BOX_SHADOW_T_T}
+            py="2rem"
+            px="4rem"
+          >
+            <UnsavedChangesWarningContent
+              onDiscard={() => {
+                // if (discardEdits) discardEdits();
+                blocker.proceed();
+              }}
+              onKeepEditing={() => blocker.reset()}
+            />
+          </DrawerContent>
+        </Drawer>
+        // </Hide>
+      )}
+      <FieldArray name="hats">
+        {({ remove, push }) => (
+          <>
+            <Show below="md">
+              <Portal>
+                <Box
+                  position="fixed"
+                  top={0}
+                  h="100vh"
+                  w="full"
+                  bg="neutral-1"
+                  px="1rem"
+                  pt={headerHeight}
+                  overflow="scroll"
                 >
                   <Flex
-                    gap="0.5rem"
-                    alignItems="center"
-                    aria-label={t('editRoles')}
-                    onClick={() => {
-                      if (hatIndex === -1) {
-                        remove(hatIndex);
-                      }
-                      navigate(DAO_ROUTES.rolesEdit.relative(addressPrefix, daoAddress));
-                    }}
-                  >
-                    <Icon
-                      as={ArrowLeft}
-                      boxSize="1.5rem"
-                    />
-                    <Text textStyle="display-lg">{t('editRoles')}</Text>
-                  </Flex>
-                  <Box position="relative">
-                    <EditRoleMenu
-                      onRemove={() => {
-                        setFieldValue('hasSavedEdits', true);
-                        navigate(DAO_ROUTES.rolesEdit.relative(addressPrefix, daoAddress));
-                      }}
-                      hatId={hatEditingId}
-                    />
-                  </Box>
-                </Flex>
-                <Box pb="5rem">
-                  <RoleFormTabs
-                    hatId={hatEditingId}
-                    push={push}
-                  />
-                </Box>
-              </Box>
-            </Portal>
-          </Show>
-          <Show above="md">
-            <Drawer
-              isOpen
-              placement="right"
-              onClose={() => {
-                if (hatIndex === -1) {
-                  remove(hatIndex);
-                }
-                navigate(DAO_ROUTES.rolesEdit.relative(addressPrefix, daoAddress));
-              }}
-            >
-              <DrawerContent
-                minW="50%"
-                bg="neutral-2"
-                pt="1rem"
-              >
-                <DrawerBody h="100vh">
-                  <Flex
                     justifyContent="space-between"
-                    my="1rem"
+                    alignItems="center"
+                    my="1.75rem"
                   >
                     <Flex
-                      gap="1rem"
+                      gap="0.5rem"
                       alignItems="center"
+                      aria-label={t('editRoles')}
+                      onClick={() => {
+                        if (hatIndex === -1) {
+                          remove(hatIndex);
+                        }
+                        goBackToRolesEdit();
+                      }}
                     >
-                      <IconButton
-                        variant="tertiary"
-                        size="icon-sm"
-                        aria-label="Close Drawer"
-                        as={X}
-                        onClick={() => {
-                          navigate(DAO_ROUTES.rolesEdit.relative(addressPrefix, daoAddress));
-                        }}
+                      <Icon
+                        as={ArrowLeft}
+                        boxSize="1.5rem"
                       />
-                      <Text
-                        textStyle="body-base"
-                        color="white-0"
-                      >
-                        {t('editRole')}
-                      </Text>
+                      <Text textStyle="display-lg">{t('editRoles')}</Text>
                     </Flex>
                     <Box position="relative">
                       <EditRoleMenu
-                        hatId={hatEditingId}
                         onRemove={() => {
-                          navigate(DAO_ROUTES.rolesEdit.relative(addressPrefix, daoAddress));
+                          setFieldValue('hasSavedEdits', true);
+                          goBackToRolesEdit();
                         }}
+                        hatId={hatEditingId}
                       />
                     </Box>
                   </Flex>
-                  <RoleFormTabs
-                    hatId={hatEditingId}
-                    push={push}
-                  />
-                </DrawerBody>
-              </DrawerContent>
-            </Drawer>
-          </Show>
-        </>
-      )}
-    </FieldArray>
+                  <Box pb="5rem">
+                    <RoleFormTabs
+                      hatId={hatEditingId}
+                      push={push}
+                    />
+                  </Box>
+                </Box>
+              </Portal>
+            </Show>
+            <Show above="md">
+              <Drawer
+                isOpen
+                placement="right"
+                onClose={() => {
+                  if (hatIndex === -1) {
+                    remove(hatIndex);
+                  }
+                  goBackToRolesEdit();
+                }}
+              >
+                <DrawerContent
+                  minW="50%"
+                  bg="neutral-2"
+                  pt="1rem"
+                >
+                  <DrawerBody h="100vh">
+                    <Flex
+                      justifyContent="space-between"
+                      my="1rem"
+                    >
+                      <Flex
+                        gap="1rem"
+                        alignItems="center"
+                      >
+                        <IconButton
+                          variant="tertiary"
+                          size="icon-sm"
+                          aria-label="Close Drawer"
+                          as={X}
+                          onClick={() => {
+                            goBackToRolesEdit();
+                          }}
+                        />
+                        <Text
+                          textStyle="body-base"
+                          color="white-0"
+                        >
+                          {t('editRole')}
+                        </Text>
+                      </Flex>
+                      <Box position="relative">
+                        <EditRoleMenu
+                          hatId={hatEditingId}
+                          onRemove={() => {
+                            navigate(DAO_ROUTES.rolesEdit.relative(addressPrefix, daoAddress));
+                          }}
+                        />
+                      </Box>
+                    </Flex>
+                    <RoleFormTabs
+                      hatId={hatEditingId}
+                      push={push}
+                    />
+                  </DrawerBody>
+                </DrawerContent>
+              </Drawer>
+            </Show>
+          </>
+        )}
+      </FieldArray>
+    </>
   );
 }
